reduced unnecessary complexity
