+++ src/cli/cardano/cli/cli.rs
@@ -1,4 +1,4 @@
-use crate::cli::utils::Terminal;
+use crate::cli::utils::print;
 use anyhow::Result;
 use console::Emoji;
 use structopt::StructOpt;
@@ -8,12 +8,12 @@ pub enum CliCommand {}
 
 impl CliCommand {
     pub fn install_cli() -> Result<()> {
-        Terminal::print("", "Installing the latest cardano-cli", Emoji("🤟", ""))?;
+        print("", "Installing the latest cardano-cli", Emoji("🤟", ""))?;
         Ok(())
     }
 
     pub fn uninstall_cli() -> Result<()> {
-        Terminal::print("white", "Uninstalling cardano-cli", Emoji("💔", ""))?;
+        print("white", "Uninstalling cardano-cli", Emoji("💔", ""))?;
         Ok(())
     }
 }
