+++ src/cli/cardano/wallet/wallet.rs
@@ -1,4 +1,4 @@
-use crate::cli::utils::Terminal;
+use crate::cli::utils::print;
 use anyhow::Result;
 use console::Emoji;
 use structopt::StructOpt;
@@ -8,12 +8,12 @@ pub enum WalletCommand {}
 
 impl WalletCommand {
     pub async fn install_wallet() -> Result<()> {
-        Terminal::print("white", "Installing cardano-wallet", Emoji("🤟", ""))?;
+        print("white", "Installing cardano-wallet", Emoji("🤟", ""))?;
         Ok(())
     }
 
     pub async fn uninstall_wallet() -> Result<()> {
-        Terminal::print("white", "Uninstalling cardano-wallet", Emoji("💔", ""))?;
+        print("white", "Uninstalling cardano-wallet", Emoji("💔", ""))?;
         Ok(())
     }
 }
