+++ src/cli/utils/mod.rs
@@ -1,2 +1,2 @@
-pub mod terminal;
-pub use terminal::*;
+pub mod utils;
+pub use utils::*;
